\name{gcTensor-package}
\alias{gcTensor-package}
\alias{gcTensor}
\docType{package}
\title{
\packageTitle{gcTensor}
}
\description{
\packageDescription{gcTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gcTensor}
\packageIndices{gcTensor}
}
\author{
\packageAuthor{gcTensor}

Maintainer: \packageMaintainer{gcTensor}
}
\references{
Y. Kenan Yilmaz, et. al., (2011). Generalised Coupled Tensor Factorisation,  NIPS

Beyza Ermis, et. al., (2015). Link prediction in heterogeneous data via generalized coupled tensor factorization, Data Mining and Knowledge Discovery
}

\keyword{ package }
\examples{
ls("package:gcTensor")
}